backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
sqlmodel==0.0.22
python-multipart==0.0.9
bcrypt==4.2.0
itsdangerous==2.2.0
httpx==0.27.2
Jinja2==3.1.4
PyJWT==2.9.0
langchain==0.2.12
langchain-core==0.2.36
langchain-openai==0.1.22
langgraph==0.2.15
openai==1.48.0
tiktoken==0.7.0
pypdf==5.0.0
python-docx==1.1.2
pandas==2.2.2
xlrd==2.0.1
Authlib==1.3.2
python-pptx==1.0.2
pytesseract==0.3.13
stripe==10.12.0
beautifulsoup4==4.12.3
